found partialy why the index doesnt work
if ep is small (top layer generated early) the results are good. If ep is big (top layer generated late) the results are bad. Can fix by creating many layers at first insert (ep=0), but if new layer appears on top, results are bad. Have to work on that, but otherwise working fine