changed how I compute bf data
before, I just passed a process number to the same binary and let it compute that split of the data. Parallelization was achived through the launch of multiple programs with different positional parameters. I changed the script so that is uses Rusts concurrency features. It was surprisingly easy to set-up. Now I only need to pass the number of threads I want to spawn, and the rest is done by the binary.