turns out naive filtering approach was good
I computed bf NNs and retained only words that began withe the letter e in the neighbors list. The index, filtering on this condition on the layer 0, yields very high recall, but with a very significant increase in latancy. I guess its natural, the filtering is very agressive (most words dont beging with e)