have something that starts to work
the partitioning works well, each thread can add its nodes without sync. But I still insert frontier nodes in a single thread, while I could re-partition like in the first step. Also, something should be done regarding the fact that we have to wait for the biggest thread to finish before passing to the frontier nodes, leaving the rest of the cores idle for a long time